ENH: rename n6 collector to stomp
fixes certtools/intelmq#716
