+++ intelmq/bots/collectors/stomp/REQUIREMENTS.txt
@@ -1,0 +1,1 @@
+stomp.py>=4.1.8
