+++ intelmq/tests/bots/collectors/stomp/test_collector.py
